+++ components/Payment/Icons/PSP.js
@@ -1,0 +1,33 @@
+import React from 'react'
+
+export const Visa = () => (
+  <svg width='51' height='40' viewBox='0 -4 51 36'>
+    <path fill='#FFF' d='M.493 31.56h49.7V.254H.493' />
+    <path fill='#004F8B' d='M18.65 23.845l2.294-16.342h3.67l-2.297 16.342H18.65M35.624 7.905c-.727-.33-1.866-.686-3.29-.686-3.624 0-6.177 2.215-6.2 5.392-.02 2.348 1.824 3.658 3.216 4.44 1.428.8 1.908 1.312 1.9 2.027-.008 1.094-1.14 1.595-2.194 1.595-1.468 0-2.248-.248-3.453-.858l-.473-.26-.515 3.66c.857.455 2.442.85 4.087.87 3.856 0 6.36-2.19 6.388-5.583.014-1.86-.963-3.274-3.08-4.44-1.282-.756-2.068-1.26-2.06-2.026 0-.68.665-1.406 2.102-1.406 1.2-.022 2.068.295 2.746.626l.328.19.498-3.543M45.03 7.518h-2.834c-.878 0-1.535.29-1.92 1.355l-5.45 14.978h3.853s.63-2.013.772-2.455l4.698.007c.11.572.447 2.45.447 2.45H48L45.03 7.517M40.51 18.052c.303-.942 1.46-4.57 1.46-4.57-.02.045.302-.945.488-1.558l.248 1.41.85 4.718h-3.046M15.573 7.514L11.98 18.66l-.38-2.266c-.67-2.61-2.753-5.44-5.083-6.855l3.285 14.29 3.88-.004L19.46 7.514h-3.887' />
+    <path d='M8.65 7.504H2.732l-.047.34c4.603 1.353 7.648 4.622 8.913 8.55l-1.288-7.51C10.09 7.85 9.446 7.54 8.65 7.504' fill='#EE9F2D' />
+  </svg>
+)
+
+export const Mastercard = () => (
+  <svg width='60' height='40' viewBox='0 -1 60 39'>
+    <path d='M22.937 18.437c0-5.855 2.746-11.07 7.02-14.443C26.808 1.506 22.824.017 18.49.017 8.277.017 0 8.264 0 18.437c0 10.173 8.278 18.42 18.49 18.42 4.333 0 8.316-1.49 11.468-3.977-4.275-3.374-7.02-8.587-7.02-14.443' fill='#D9252C' />
+    <path d='M22.937 18.437c0 5.856 2.746 11.07 7.02 14.443 4.276-3.374 7.022-8.587 7.022-14.443 0-5.855-2.747-11.07-7.022-14.443-4.275 3.374-7.02 8.588-7.02 14.443' fill='#E26229' />
+    <path d='M41.427.017c-4.334 0-8.316 1.49-11.47 3.977 4.276 3.374 7.022 8.588 7.022 14.443 0 5.856-2.747 11.07-7.022 14.443 3.153 2.488 7.135 3.977 11.47 3.977 10.21 0 18.488-8.247 18.488-18.42 0-10.173-8.278-18.42-18.49-18.42' fill='#ED982E' />
+  </svg>
+)
+
+export const Postcard = () => (
+  <svg width='79' height='40' viewBox='0 -6 79 28'>
+    <path d='M3.68 3.543v4.26h.98c1.4 0 2.66-.46 2.66-2.08s-1.22-2.18-2.66-2.18h-.98zm-2.8 11.76V1.343h3.54c3.04 0 5.82.88 5.82 4.22 0 3.26-2.38 4.44-5.38 4.44H3.68v5.3H.88zm10.46-5.08c0-3.4 2.34-5.46 5.38-5.46 3.06 0 5.4 2.06 5.4 5.46 0 2.96-1.98 5.32-5.4 5.32-3.4 0-5.38-2.36-5.38-5.32zm2.8-.32c0 1.7.54 3.56 2.58 3.56 2.06 0 2.6-1.84 2.6-3.56 0-1.58-.84-3.06-2.6-3.06-1.74 0-2.58 1.52-2.58 3.06zm16.38-4.76c-.84-.18-1.66-.38-2.92-.38-2.1 0-4.04.98-4.04 3.2 0 3.72 4.66 2.34 4.66 4.36 0 1-1.08 1.26-1.76 1.26-.92 0-1.8-.26-2.64-.68l-.16 2.2c.98.26 2.02.44 3.04.44 2.18 0 4.32-.96 4.32-3.36 0-3.72-4.66-2.68-4.66-4.32 0-.88.86-1.14 1.58-1.14.98 0 1.58.16 2.4.44l.18-2.02zm3.46 7.06c0 2.04 1.16 3.34 3.26 3.34.8 0 1.42-.08 1.98-.24l-.08-2.12c-.28.18-.74.28-1.22.28-1.02 0-1.26-.82-1.26-1.7v-4.8h2.38v-1.96h-2.38v-2.92l-2.68.86v2.06H32v1.96h1.98v5.24zm13.82-7.1c-.74-.22-1.52-.34-2.6-.34-3.22 0-4.98 2.36-4.98 5.34 0 3.14 1.7 5.44 5.08 5.44 1.1 0 1.9-.1 2.68-.34l-.12-2.22c-.46.2-1.3.48-2.06.48-1.88 0-2.78-1.46-2.78-3.36 0-1.94 1.06-3.26 2.72-3.26.82 0 1.34.16 1.82.38l.24-2.12zm2.5 2.74l-.08-2.3c1.04-.48 2.44-.78 3.58-.78 3.14 0 4.44 1.3 4.44 4.34v1.32c0 1.04.02 1.82.04 2.58.02.78.06 1.5.12 2.3h-2.36c-.1-.54-.1-1.22-.12-1.54h-.04c-.62 1.14-1.96 1.78-3.18 1.78-1.82 0-3.6-1.1-3.6-3.06 0-1.54.74-2.44 1.76-2.94s2.34-.6 3.46-.6h1.48c0-1.66-.74-2.22-2.32-2.22-1.14 0-2.28.44-3.18 1.12zm1.36 4.48c0 .88.74 1.26 1.7 1.26.82 0 1.46-.36 1.88-.92.44-.58.56-1.32.56-2.12h-1.16c-1.2 0-2.98.2-2.98 1.78zm9.28 2.98h2.68v-3.78c0-1.48 0-4.2 2.18-4.2.48 0 .96.08 1.2.22v-2.7c-.28-.08-.56-.08-.82-.08-1.6 0-2.7 1.62-2.82 2.58h-.04v-2.34h-2.38v10.3zm9.86-5.18c0 1.42.5 3.34 2.26 3.34 1.78 0 2.36-1.88 2.36-3.34 0-1.44-.68-3.28-2.4-3.28-1.74 0-2.22 1.9-2.22 3.28zm4.72 5.18v-1.16h-.04c-.8 1-1.9 1.4-3.16 1.4-3.16 0-4.32-2.6-4.32-5.42 0-2.84 1.16-5.36 4.32-5.36 1.34 0 2.28.46 3.08 1.44h.04v-5.9h2.68v15h-2.6z' fill='#000' />
+  </svg>
+)
+
+export const PayPal = () => (
+  <svg width='77' height='40' viewBox='0 -9 77 31'>
+    <path d='M52.602 5.675c-.328 2.148-1.968 2.148-3.555 2.148h-.903l.634-4.01c.038-.243.247-.422.493-.422h.414c1.08 0 2.1 0 2.627.617.315.368.41.913.292 1.668zM51.912.07h-5.985c-.41 0-.757.298-.82.703l-2.42 15.343c-.05.303.185.577.492.577h3.07c.285 0 .53-.21.574-.49l.686-4.35c.064-.406.412-.703.82-.703h1.895c3.94 0 6.216-1.908 6.81-5.69.268-1.652.01-2.95-.763-3.86-.85-1-2.36-1.53-4.36-1.53z' fill='#238EC2' />
+    <path d='M9.922 5.675c-.327 2.148-1.968 2.148-3.554 2.148h-.903l.633-4.01c.038-.243.247-.422.493-.422h.414c1.08 0 2.1 0 2.627.617.316.368.41.913.292 1.668zM9.232.07H3.246c-.41 0-.757.298-.82.703L.006 16.116c-.048.303.185.577.492.577h2.856c.41 0 .758-.298.82-.702l.654-4.138c.064-.405.412-.702.822-.702h1.893c3.942 0 6.216-1.908 6.81-5.69.268-1.652.012-2.95-.763-3.86-.85-1-2.36-1.53-4.36-1.53zM23.123 11.183c-.277 1.637-1.576 2.736-3.234 2.736-.832 0-1.497-.268-1.925-.774-.424-.502-.583-1.217-.45-2.013.26-1.622 1.58-2.757 3.212-2.757.814 0 1.475.27 1.91.78.44.515.612 1.235.486 2.027zm3.994-5.577H24.25c-.244 0-.453.178-.49.42l-.127.802-.2-.29c-.62-.9-2.004-1.202-3.385-1.202-3.166 0-5.87 2.4-6.396 5.764-.274 1.678.115 3.282 1.066 4.402.875 1.03 2.122 1.457 3.61 1.457 2.55 0 3.966-1.64 3.966-1.64l-.128.796c-.048.303.186.577.493.577h2.58c.408 0 .757-.298.82-.702l1.55-9.808c.047-.302-.187-.576-.493-.576z' fill='#253666' />
+    <path d='M65.803 11.183c-.277 1.637-1.576 2.736-3.234 2.736-.832 0-1.497-.268-1.925-.774-.424-.502-.584-1.217-.45-2.013.26-1.622 1.58-2.757 3.212-2.757.814 0 1.474.27 1.91.78.44.515.612 1.235.486 2.027zm3.993-5.577H66.93c-.244 0-.453.178-.492.42l-.126.802-.2-.29c-.62-.9-2.004-1.202-3.385-1.202-3.165 0-5.87 2.4-6.396 5.764-.273 1.678.116 3.282 1.068 4.402.874 1.03 2.122 1.457 3.61 1.457 2.55 0 3.966-1.64 3.966-1.64l-.128.796c-.048.303.186.577.493.577h2.58c.408 0 .756-.298.82-.702l1.55-9.808c.047-.302-.187-.576-.494-.576z' fill='#238EC2' />
+    <path d='M42.378 5.606h-2.88c-.276 0-.533.137-.688.364l-3.973 5.852L33.154 6.2c-.106-.353-.43-.594-.797-.594h-2.83c-.343 0-.583.336-.473.66l3.17 9.308-2.982 4.21c-.234.33.002.786.407.786h2.876c.273 0 .53-.133.684-.357L42.788 6.39c.23-.332-.007-.784-.41-.784' fill='#253666' />
+    <path d='M73.174.492l-2.456 15.625c-.048.302.186.576.493.576h2.47c.41 0 .758-.298.822-.702L76.924.648C76.97.344 76.737.07 76.43.07h-2.763c-.246 0-.455.18-.493.422' fill='#238EC2' />
+  </svg>
+)
