+++ components/Pledge/Form.js
@@ -70,6 +70,7 @@ class Pledge extends Component {
     const userPrice = !!query.userPrice
 
     return {
+      paymentMethods: pkg ? pkg.paymentMethods : undefined,
       total: values.price || undefined,
       user: {
         firstName: values.firstName,
@@ -303,6 +304,7 @@ query pledgeForm($crowdfundingName: String!) {
     packages {
       id
       name
+      paymentMethods
       options {
         id
         price
