+++ components/Pledge/Submit.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { css } from 'glamor'
 import { graphql, compose } from 'react-apollo'
 import gql from 'graphql-tag'
 
@@ -13,74 +12,27 @@ import { meQuery } from '../../lib/apollo/withMe'
 import { chfFormat } from '../../lib/utils/format'
 import track from '../../lib/piwik'
 
-import FieldSet from '../FieldSet'
-
-import * as postfinance from './postfinance'
-import * as paypal from './paypal'
-import loadStripe from './stripe'
 import { gotoMerci } from './Merci'
 
-import AddressForm, { COUNTRIES } from '../Account/AddressForm'
+import { COUNTRIES } from '../Account/AddressForm'
 import { query as addressQuery } from '../Account/UpdateMe'
 
-import LockIcon from './Icons/Lock'
-import * as PSPIcons from './Icons/PSP'
+import FieldSet from '../FieldSet'
 
 import {
-  PUBLIC_BASE_URL,
-  PF_FORM_ACTION,
-  PAYPAL_FORM_ACTION
+  PUBLIC_BASE_URL
 } from '../../lib/constants'
 
 import {
-  Interaction, Button, Label, Checkbox,
-  colors, fontFamilies, InlineSpinner,
+  Interaction, Button, Checkbox,
+  colors, InlineSpinner,
   RawHtml
 } from '@project-r/styleguide'
 
-const {H2, P} = Interaction
+import PaymentForm from '../Payment/Form'
+import loadStripe from '../Payment/stripe'
 
-const PAYMENT_METHODS = [
-  {
-    disabled: false,
-    key: 'STRIPE',
-    Icon: ({state: {values, stripe}}) => {
-      let cardType = null
-      if (stripe && values && values.cardNumber) {
-        cardType = stripe.card.cardType(values.cardNumber)
-        if (cardType === 'Unknown') {
-          cardType = null
-        }
-      }
-      return (
-        <span>
-          <span style={{opacity: !cardType || cardType === 'Visa' ? 1 : 0.4}}>
-            <PSPIcons.Visa />
-          </span>
-          <span style={{display: 'inline-block', width: 10}} />
-          <span style={{opacity: !cardType || cardType === 'MasterCard' ? 1 : 0.4}}>
-            <PSPIcons.Mastercard />
-          </span>
-        </span>
-      )
-    }
-  },
-  {
-    disabled: false,
-    key: 'POSTFINANCECARD',
-    bgColor: '#FCCC12',
-    Icon: PSPIcons.Postcard
-  },
-  {
-    disabled: false,
-    key: 'PAYMENTSLIP'
-  },
-  {
-    disabled: false,
-    key: 'PAYPAL',
-    Icon: PSPIcons.PayPal
-  }
-]
+const {P} = Interaction
 
 const objectValues = (object) => Object.keys(object).map(key => object[key])
 const simpleHash = (object, delimiter = '|') => {
@@ -92,50 +44,10 @@ const simpleHash = (object, delimiter = '|') => {
   }).join(delimiter)
 }
 
-const styles = {
-  secure: css({
-    fontFamily: fontFamilies.sansSerifMedium,
-    fontSize: 14,
-    color: colors.primary,
-    marginBottom: 20,
-    marginTop: 10,
-    '& svg': {
-      marginRight: 5
-    }
-  }),
-  paymentMethod: css({
-    fontFamily: fontFamilies.sansSerifMedium,
-    fontSize: 14,
-    color: '#000',
-    display: 'inline-block',
-    border: `1px solid ${colors.secondary}`,
-    padding: 10,
-    cursor: 'pointer',
-    marginRight: 10,
-    marginBottom: 10,
-    lineHeight: 0,
-    height: 62,
-    verticalAlign: 'top',
-
-    '& input': {
-      display: 'none'
-    }
-  }),
-  paymentMethodTextOnly: css({
-    lineHeight: '40px'
-  }),
-  paymentMethodTextWithIcon: css({
-    position: 'absolute',
-    left: -10000,
-    top: 'auto'
-  })
-}
-
 class Submit extends Component {
   constructor (props) {
     super(props)
     this.state = {
-      paymentMethod: 'STRIPE',
       emailVerify: false,
       legal: false,
       values: {
@@ -158,14 +70,8 @@ class Submit extends Component {
       this.state.pledgeHash = hash
       this.state.pledgeId = props.basePledge.id
     }
-    this.amountRefSetter = (ref) => {
-      this.amountRef = ref
-    }
-    this.postFinanceFormRef = (ref) => {
-      this.postFinanceForm = ref
-    }
-    this.payPalFormRef = (ref) => {
-      this.payPalForm = ref
+    this.paymentRef = (ref) => {
+      this.payment = ref
     }
   }
   componentWillReceiveProps (nextProps) {
@@ -236,7 +142,7 @@ class Submit extends Component {
       // - this can be missing if returning from a PSP redirect
       // - in those cases we create a new pledge
       (
-        this.state.paymentMethod !== 'POSTFINANCECARD' ||
+        this.state.values.paymentMethod !== 'POSTFINANCECARD' ||
         this.state.pledgeResponse
       )
     ) {
@@ -280,7 +186,7 @@ class Submit extends Component {
       })
   }
   payPledge (pledgeId, pledgeResponse) {
-    const {paymentMethod} = this.state
+    const { paymentMethod } = this.state.values
 
     if (paymentMethod === 'PAYMENTSLIP') {
       this.payWithPaymentSlip(pledgeId)
@@ -365,110 +271,119 @@ class Submit extends Component {
       })
   }
   payWithStripe (pledgeId) {
-    const {t, total} = this.props
-    const {values, stripe} = this.state
+    const { t, total } = this.props
+    const { values } = this.state
 
     this.setState(() => ({
       loading: t('pledge/submit/loading/stripe')
     }))
-    stripe.source.create({
-      type: 'card',
-      currency: 'CHF',
-      amount: total,
-      usage: 'reusable',
-      card: {
-        number: values.cardNumber,
-        cvc: values.cardCVC,
-        exp_month: values.cardMonth,
-        exp_year: values.cardYear
-      },
-      metadata: {
-        pledgeId
-      }
-    }, (status, source) => {
-      if (status !== 200) {
-        // source.error.type
-        // source.error.param
-        // source.error.message
-        // see https://stripe.com/docs/api#errors
-        // - never happens because we use client validation
-        // - only when charging some additional errors can happen
-        //   those are handled server side in the pay mutation
-        // - if it happens, we simply display the English message
-        // test cards https://stripe.com/docs/testing#cards
-        this.setState(() => ({
-          loading: false,
-          paymentError: source.error.message
-        }))
-        return
-      }
-      this.setState({
-        loading: false,
-        paymentError: undefined
-      })
 
-      if (source.card.three_d_secure === 'required') {
-        this.setState(() => ({
-          loading: t('pledge/submit/loading/stripe/3dsecure')
-        }))
-        stripe.source.create({
-          type: 'three_d_secure',
-          currency: 'CHF',
-          amount: total,
-          three_d_secure: {
-            card: source.id
-          },
-          redirect: {
-            return_url: `${PUBLIC_BASE_URL}/pledge?pledgeId=${pledgeId}&stripe=1`
-          },
-          metadata: {
-            pledgeId
-          }
-        }, (status, source3d) => {
-          if (status !== 200) {
-            this.setState(() => ({
-              loading: false,
-              paymentError: t.first([
-                `pledge/3dsecure/${source3d.error.code}`,
-                'pledge/3dsecure/unkown'
-              ])
-            }))
-            return
-          }
-          if (source3d.redirect.status === 'succeeded') {
-            // can charge immediately
-            this.pay({
-              pledgeId,
-              method: 'STRIPE',
-              sourceId: source3d.id,
-              pspPayload: JSON.stringify(source3d)
-            })
-          } else if (source3d.redirect.status === 'failed') {
-            // no support or bank 3D Secure down
-            this.setState(() => ({
-              loading: false,
-              paymentError: t('pledge/3dsecure/redirect/failed')
-            }))
-          } else {
-            window.location = source3d.redirect.url
-          }
+    loadStripe().then(stripe => {
+      stripe.source.create({
+        type: 'card',
+        currency: 'CHF',
+        amount: total,
+        usage: 'reusable',
+        card: {
+          number: values.cardNumber,
+          cvc: values.cardCVC,
+          exp_month: values.cardMonth,
+          exp_year: values.cardYear
+        },
+        metadata: {
+          pledgeId
+        }
+      }, (status, source) => {
+        if (status !== 200) {
+          // source.error.type
+          // source.error.param
+          // source.error.message
+          // see https://stripe.com/docs/api#errors
+          // - never happens because we use client validation
+          // - only when charging some additional errors can happen
+          //   those are handled server side in the pay mutation
+          // - if it happens, we simply display the English message
+          // test cards https://stripe.com/docs/testing#cards
+          this.setState(() => ({
+            loading: false,
+            paymentError: source.error.message
+          }))
+          return
+        }
+        this.setState({
+          loading: false,
+          paymentError: undefined
         })
 
-        return
-      }
+        if (source.card.three_d_secure === 'required') {
+          this.setState(() => ({
+            loading: t('pledge/submit/loading/stripe/3dsecure')
+          }))
+          stripe.source.create({
+            type: 'three_d_secure',
+            currency: 'CHF',
+            amount: total,
+            three_d_secure: {
+              card: source.id
+            },
+            redirect: {
+              return_url: `${PUBLIC_BASE_URL}/pledge?pledgeId=${pledgeId}&stripe=1`
+            },
+            metadata: {
+              pledgeId
+            }
+          }, (status, source3d) => {
+            if (status !== 200) {
+              this.setState(() => ({
+                loading: false,
+                paymentError: t.first([
+                  `pledge/3dsecure/${source3d.error.code}`,
+                  'pledge/3dsecure/unkown'
+                ])
+              }))
+              return
+            }
+            if (source3d.redirect.status === 'succeeded') {
+              // can charge immediately
+              this.pay({
+                pledgeId,
+                method: 'STRIPE',
+                sourceId: source3d.id,
+                pspPayload: JSON.stringify(source3d)
+              })
+            } else if (source3d.redirect.status === 'failed') {
+              // no support or bank 3D Secure down
+              this.setState(() => ({
+                loading: false,
+                paymentError: t('pledge/3dsecure/redirect/failed')
+              }))
+            } else {
+              window.location = source3d.redirect.url
+            }
+          })
 
-      this.pay({
-        pledgeId,
-        method: 'STRIPE',
-        sourceId: source.id,
-        pspPayload: JSON.stringify(source)
+          return
+        }
+
+        this.pay({
+          pledgeId,
+          method: 'STRIPE',
+          sourceId: source.id,
+          pspPayload: JSON.stringify(source)
+        })
       })
+    })
+    .catch(() => {
+      this.setState(() => ({
+        loading: false,
+        paymentError: t('pledge/submit/stripe/js/failed')
+      }))
     })
   }
   getErrorMessages () {
     const {
-      paymentMethod, loadingStripe, loadingStripeError,
-      legal
+      legal,
+      values
     } = this.state
     const {t, options} = this.props
 
@@ -478,9 +393,7 @@ class Submit extends Component {
       .concat(objectValues(this.props.errors))
       .concat(objectValues(this.state.errors))
       .concat([
-        !paymentMethod && t('pledge/submit/payMethod/error'),
-        paymentMethod === 'STRIPE' && loadingStripe && t('pledge/submit/stripe/js/loading'),
-        paymentMethod === 'STRIPE' && loadingStripeError && t('pledge/submit/stripe/js/failed'),
+        !values.paymentMethod && t('pledge/submit/payMethod/error'),
         !legal && t('pledge/submit/legal/error')
       ])
       .filter(Boolean)
@@ -488,247 +401,46 @@ class Submit extends Component {
   render () {
     const {
       emailVerify,
-      paymentMethod,
       paymentError,
       submitError,
       signInError,
       loading
     } = this.state
-    const {me, user, t} = this.props
+    const {
+      me, user, t,
+      paymentMethods
+    } = this.props
 
     const errorMessages = this.getErrorMessages()
 
     return (
       <div>
-        <H2>{t('pledge/submit/payMethod/title')}</H2>
-        <div {...styles.secure}>
-          <LockIcon /> {t('pledge/submit/secure')}
-        </div>
-        <P>
-          <Label>{t('pledge/submit/payMethod/label')}</Label><br />
-          {PAYMENT_METHODS.filter(pm => !pm.disabled).map((pm) => (
-            <label key={pm.key}
-              {...styles.paymentMethod}
-              style={{
-                backgroundColor: pm.bgColor,
-                opacity: paymentMethod === pm.key ? 1 : 0.4
-              }}>
-              <input
-                type='radio'
-                name='paymentMethod'
-                disabled={pm.disabled}
-                onChange={(event) => {
-                  event.preventDefault()
-                  const value = event.target.value
-                  this.setState((state) => {
-                    const next = {
-                      showErrors: false,
-                      paymentMethod: value
-                    }
-                    if (value !== state.paymentMethod) {
-                      next.errors = {}
-                    }
+        <PaymentForm
+          ref={this.paymentRef}
+          t={t}
+          payload={{
+            id: this.state.pledgeId,
+            userId: this.state.userId,
+            total: this.props.total,
+            pfAliasId: this.state.pfAliasId,
+            pfSHA: this.state.pfSHA
+          }}
+          allowedMethods={paymentMethods}
+          onChange={fields => {
+            this.setState(state => {
+              const nextState = FieldSet.utils.mergeFields(fields)(state)
 
-                    return next
-                  })
-                }}
-                value={pm.key}
-                checked={paymentMethod === pm.key} />
-              {pm.Icon ? <pm.Icon state={this.state} /> : null}
-              <span {...(pm.Icon
-                ? styles.paymentMethodTextWithIcon
-                : styles.paymentMethodTextOnly
-              )}>
-                {t(`pledge/submit/pay/method/${pm.key}`)}
-              </span>
-            </label>
-          ))}
-        </P>
+              if (state.values.paymentMethod !== nextState.values.paymentMethod) {
+                nextState.showErrors = false
+                nextState.errors = {}
+              }
 
-        {(paymentMethod === 'PAYMENTSLIP') && (
-          <div>
-            <Label>
-              {t('pledge/submit/paymentslip/explanation')}
-            </Label><br /><br />
-            <Label>{t('pledge/submit/paymentslip/title')}</Label>
-            <AddressForm
-              values={this.state.values}
-              errors={this.state.errors}
-              dirty={this.state.dirty}
-              onChange={(fields) => {
-                this.setState(FieldSet.utils.mergeFields(fields))
-              }} />
-            <br />
-            { /* <div style={{marginBottom: 5}}>
-              <Radio
-                checked={!this.state.values.paperInvoice}
-                onChange={() => {
-                  this.setState((state) => ({
-                    values: {
-                      ...state.values,
-                      paperInvoice: false
-                    }
-                  }))
-                }}>
-                {t('pledge/submit/paymentslip/emailInvoice')}
-              </Radio>
-            </div>
-            <div>
-              <Radio
-                checked={!!this.state.values.paperInvoice}
-                onChange={() => {
-                  this.setState((state) => ({
-                    values: {
-                      ...state.values,
-                      paperInvoice: true
-                    }
-                  }))
-                }}>
-                {t('pledge/submit/paymentslip/paperInvoice')}
-              </Radio>
-            </div> */ }
-            <br />
-          </div>
-        )}
-        {(paymentMethod === 'STRIPE') && (
-          <from method='post' onSubmit={(e) => {
-            e.preventDefault()
-          }}>
-            <Label>{t('pledge/submit/stripe/prefered')}</Label>
-            <FieldSet
-              values={this.state.values}
-              errors={this.state.errors}
-              dirty={this.state.dirty}
-              fields={[
-                {
-                  label: t('pledge/submit/stripe/card/label'),
-                  name: 'cardNumber',
-                  autoComplete: 'cc-number',
-                  mask: '1111 1111 1111 1111',
-                  validator: (value) => (
-                    (
-                      !value &&
-                      t('pledge/submit/stripe/card/error/empty')
-                    ) || (
-                      !!this.state.stripe &&
-                      !this.state.stripe.card.validateCardNumber(value) &&
-                      t('pledge/submit/stripe/card/error/invalid')
-                    )
-                  )
-                },
-                {
-                  label: t('pledge/submit/stripe/month/label'),
-                  name: 'cardMonth',
-                  autoComplete: 'cc-exp-month'
-                },
-                {
-                  label: t('pledge/submit/stripe/year/label'),
-                  name: 'cardYear',
-                  autoComplete: 'cc-exp-year'
-                },
-                {
-                  label: t('pledge/submit/stripe/cvc/label'),
-                  name: 'cardCVC',
-                  autoComplete: 'cc-csc',
-                  validator: (value) => (
-                    (
-                      !value &&
-                      t('pledge/submit/stripe/cvc/error/empty')
-                    ) || (
-                      !!this.state.stripe &&
-                      !this.state.stripe.card.validateCVC(value) &&
-                      t('pledge/submit/stripe/cvc/error/invalid')
-                    )
-                  )
-                }
-              ]}
-              onChange={(fields, mounting) => {
-                if (
-                  !mounting &&
-                  !this.state.stripe &&
-                  !this.state.loadingStripe
-                ) {
-                  this.setState(() => ({
-                    loadingStripe: true,
-                    loadingStripeError: undefined
-                  }))
-                  loadStripe()
-                    .then(stripe => {
-                      this.setState(() => ({
-                        loadingStripe: false,
-                        stripe
-                      }))
-                    })
-                    .catch(() => {
-                      this.setState(() => ({
-                        loadingStripe: false,
-                        loadingStripeError: t('pledge/submit/stripe/js/failed')
-                      }))
-                    })
-                }
-                this.setState((state) => {
-                  const nextState = FieldSet.utils.mergeFields(fields)(state)
-
-                  const month = nextState.values.cardMonth
-                  const year = nextState.values.cardYear
-
-                  if (
-                    year && month &&
-                    nextState.dirty.cardMonth &&
-                    nextState.dirty.cardYear &&
-                    !!this.state.stripe &&
-                    !this.state.stripe.card.validateExpiry(month, year)
-                  ) {
-                    nextState.errors.cardMonth = t('pledge/submit/stripe/month/error/invalid')
-                    nextState.errors.cardYear = t('pledge/submit/stripe/year/error/invalid')
-                  } else {
-                    nextState.errors.cardMonth = (
-                      !month && t('pledge/submit/stripe/month/error/empty')
-                    )
-                    nextState.errors.cardYear = (
-                      !year && t('pledge/submit/stripe/year/error/empty')
-                    )
-                  }
-
-                  return nextState
-                })
-              }} />
-            <br /><br />
-          </from>
-        )}
-        {(paymentMethod === 'POSTFINANCECARD') && (
-          <form ref={this.postFinanceFormRef} method='post' action={PF_FORM_ACTION}>
-            {
-              postfinance.getParams({
-                userId: this.state.userId,
-                orderId: this.state.pledgeId,
-                amount: this.props.total,
-                alias: this.state.pfAliasId,
-                sha: this.state.pfSHA
-              }).map(param => (
-                <input key={param.key}
-                  type='hidden'
-                  name={param.key}
-                  value={param.value} />
-              ))
-            }
-          </form>
-        )}
-        {(paymentMethod === 'PAYPAL') && (
-          <form ref={this.payPalFormRef} method='post' action={PAYPAL_FORM_ACTION}>
-            {
-              paypal.getParams({
-                itemName: this.state.pledgeId,
-                amount: this.props.total
-              }).map(param => (
-                <input key={param.key}
-                  type='hidden'
-                  name={param.key}
-                  value={param.value} />
-              ))
-            }
-          </form>
-        )}
+              return nextState
+            })
+          }}
+          values={this.state.values}
+          errors={this.state.errors}
+          dirty={this.state.dirty} />
         <br /><br />
         {(emailVerify && !me) && (
           <div style={{marginBottom: 40}}>
@@ -809,6 +521,7 @@ class Submit extends Component {
 Submit.propTypes = {
   me: PropTypes.object,
   user: PropTypes.object,
+  paymentMethods: PropTypes.array,
   total: PropTypes.number,
   reason: PropTypes.string,
   options: PropTypes.array.isRequired,
